Fixed generate public share link
